fix(app): simplify and refine combobox
One of the UX things that bothered me before these changes was that when you hit enter with a menu item selected, it populated the text box with the ALL CAPS version of the item and then you had to hit enter again to add the agent to the list.

I also felt like the add button was extraneous at this point but it could easily be added back in or changed to a dropdown.

Inspired by https://codesandbox.io/s/github/kentcdodds/downshift-examples?file=/src/hooks/useMultipleSelection/combobox.js:325-373